+++ CHANGELOG.md
@@ -53,4 +53,5 @@
 - MVP-356 - marketdata target-markets
 - MVP-357 - Implement new signup modal design
 - MVP-362 - styling and correct HTML for country statistics view
+- MVP-289 - add products modal to export plan
 - MVP-360 - muliple markets